+++ pyxelrest/authentication.py
@@ -31,6 +31,25 @@ def _create_authentication(service_config, open_api_security_definition, request
         logger.error('Unexpected security definition type: {0}'.format(open_api_security_definition))
 
 
+def _create_authentication_from_config(service_config, authentication_mode, authentication):
+    if 'oauth2' == authentication_mode:
+        oauth2_config = dict(service_config.oauth2)
+        return OAuth2(authorization_url=authentication.get('oauth2_auth_url'),
+                      redirect_uri_port=oauth2_config.pop('port', None),
+                      token_reception_timeout=oauth2_config.pop('timeout', None),
+                      token_reception_success_display_time=oauth2_config.pop('success_display_time', None),
+                      token_reception_failure_display_time=oauth2_config.pop('failure_display_time', None),
+                      **oauth2_config)
+    elif 'api_key' == authentication_mode:
+        if authentication.get('in') == 'query':
+            return QueryApiKey(service_config.api_key, authentication.get('name'))
+        return HeaderApiKey(service_config.api_key, authentication.get('name'))
+    elif 'basic' == authentication_mode:
+        return Basic(service_config.basic.get('username'), service_config.basic.get('password'))
+    else:
+        logger.error('Unexpected security definition type: {0}'.format(authentication_mode))
+
+
 def get_auth(udf_method, request_content):
     if not udf_method.requires_authentication(request_content):
         return None
@@ -51,6 +70,33 @@ def get_auth(udf_method, request_content):
                     authentication_modes.append(auth)
             except:
                 logger.exception('{0} authentication cannot be handled.'.format(security_definition_key))
+        # A single authentication method is required and PyxelRest support it
+        if len(authentication_modes) == 1:
+            return authentication_modes[0]
+        # Multiple authentication methods are required and PyxelRest support it
+        if len(authentication_modes) > 1:
+            return Auths(authentication_modes)
+        # Otherwise check if there is another security available
+
+    # Default to custom authentication if no security is supported
+    return ntlm_authentication
+
+
+def get_definition_retrieval_auth(service_config):
+    if not service_config.definition_retrieval_auths:
+        return None
+
+    ntlm_config = service_config.ntlm_auth if service_config.definition_retrieval_auths.pop('ntlm', None) else None
+    ntlm_authentication = NTLM(ntlm_config.get('username'), ntlm_config.get('password')) if ntlm_config else None
+
+    for authentication_mode, authentication in service_config.definition_retrieval_auths.items():
+        authentication_modes = [ntlm_authentication] if ntlm_authentication else []
+        try:
+            auth = _create_authentication_from_config(service_config, authentication_mode, authentication)
+            if auth:
+                authentication_modes.append(auth)
+        except:
+            logger.exception('{0} authentication cannot be handled.'.format(authentication_mode))
         # A single authentication method is required and PyxelRest support it
         if len(authentication_modes) == 1:
             return authentication_modes[0]